worker/requirements.txt
# Transcriptor Worker Dependencies
faster-whisper>=1.0.0
supabase>=2.4.0
pydub>=0.25.1
redis>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
